+++ Web application/packages/generic-components/components/EventDetailPage/main.js
@@ -7,9 +7,10 @@ import './style.css'
 import {Events} from 'meteor/visualisation:database'
 import {ReactiveVar} from 'meteor/reactive-var'
 
-const template = Template.EventDetailPage;
+const template = Template.EventDetailPage
+const container = Template.EventDetailPageContainer
 
-template.onCreated(function() {
+container.onCreated(function() {
 	const template = this;
 
 	this.autorun(function() {
@@ -18,26 +19,52 @@ template.onCreated(function() {
 			template.subscribe('eventsSentByModule', module._id)
 		}
 	})
+})
 
-	template.selection = new ReactiveVar(null)
+container.helpers({
+	eventsCursor: function() {
+		return Events.find({senderId: this.module._id}, {sort: {date: -1}})
+	}
+})
+
+template.onCreated(function() {
+	this.selection = new ReactiveVar(null);
+	this.stickToBottom = new ReactiveVar(true);
 })
 
 template.helpers({
-	eventsCursor: function() {
-		return Events.find({senderId: this.module._id}, {sort: {date: -1}})
-	},
 	selected(event) {
 		return event._id == Template.instance().selection.get()
 	},
-	selectedEvent() {
+	selectedEvent(cursor) {
 		const id = Template.instance().selection.get()
 		if (id == null) return undefined
-		return Events.findOne(id)
-	}
+		return cursor.collection.findOne(id)
+	},
+	stickToBottom() {return Template.instance().stickToBottom.get()}
 })
 
 template.events({
 	'click .js-item'(event, template) {
 		template.selection.set(this._id)
+	},
+	'click .js-stick'(event, template) {
+		template.stickToBottom.set(true);
+		const list = template.find('.js-eventsList');
+		list.scrollTop = list.scrollHeight;
+	},
+	'wheel .js-eventsList'(event, template) {
+		template.stickToBottom.set(false);
 	}
 })
+
+template.onRendered(function() {
+	const list = this.find('.js-eventsList');
+	const stickToBottom = this.stickToBottom;
+	list._uihooks = {
+		insertElement(node, next) {
+			$(node).insertBefore(next);
+			if (stickToBottom.get()) list.scrollTop = list.scrollHeight;
+		}
+	}
+});